# openapi/src/main/resources/api.yaml
swagger: '2.0'
info:
  description: Payment API for AIRFrench
  version: v1.0.0
  title: AIRFrench Payment
  contact:
    email: contact@airfrench.fr

host: localhost
schemes:
  - http

paths:
  /payments:
    post:
      summary: Add a new payment
      operationId: addPayment
      description: Record a new payment
      parameters:
        - in: body
          name: body
          description: Payment object to create
          schema:
            $ref: '#/definitions/Payment'
      produces:
        - application/json
      responses:
        201:
          description: Payment was created successfully
          schema:
            type: object
            properties:
              transactionId:
                type: string
                format: uuid
              date:
                type: string
                format: date-time
        400:
          description: Payment could not be created, server is either processing the payment or have already refused it
          schema:
            type: object
            properties:
              codeError:
                type: string
                pattern: '^PAYMENT_FAILED|PAYMENT_PROCESSING$'
              occurredDate:
                type: string
                format: date-time
              message:
                type: string
                format: string

  /payments/{paymentId}:
    get:
      summary: Get a payment
      operationId: getPayment
      description: Get a payment by transactionId
      parameters:
        - name: paymentId
          in: path
          description: PaymentId of the payment
          required: true
          type: string
          format: int
      produces:
        - application/json
      responses:
        200:
          description: Payment was found
          schema:
            $ref: '#/definitions/Payment'
        404:
          description: Payment was not found
          schema:
            type: object
            properties:
              message:
                type: string
                format: string

definitions:
  Payment:
    type: object
    required:
      - id
      - buyerInfo
      - sellerAccount
      - amount
      - currency
    properties:
      id:
        type: integer
        format: int64
      date:
        type: string
        format: date-time
      buyerInfo:
        type: string
        format: json
      sellerAccount:
        type: string
      amount:
        type: string
      currency:
        type: string
        format: ISO 4217 [4]